[iOS CI] Add clean to xcodebuild (#971)
* add clean command

* replace xcpretty with fastlane scan

* Fix iOS build

+++ src/Bitcoin/TransactionSigner.cpp
@@ -51,7 +51,7 @@ Result<Transaction> TransactionSigner<Transaction, TransactionBuilder>::sign() {
     tx.outputs = transaction.outputs;
     // save estimated size
     if ((input.byte_fee()) > 0 && (plan.fee > 0)) {
-        tx.previousEstimatedVirtualSize = plan.fee / input.byte_fee();
+        tx.previousEstimatedVirtualSize = static_cast<int>(plan.fee / input.byte_fee());
     }
 
     return Result<Transaction>::success(std::move(tx));
